Add bf16 coverage to LLM.int8 tests
Extends existing tests to exercise the new bf16 activation-quant path:
- test_matmullt: drop the forced fp16 cast so the existing bf16 param
  actually exercises the bf16 kernel end-to-end.
- test_int8_linear_matmul_half: parametrize over fp16/bf16.
- test_linear8bitlt_inference and new test_linear8bitlt_forward_dtypes:
  exercise the Linear8bitLt module forward in both dtypes.

Addresses review feedback on #1985.

+++ tests/test_autograd.py
@@ -76,7 +76,7 @@ def test_matmullt(
                 if not transpose[0] and not transpose[1]:
                     B2 = B2.t().contiguous()
 
-                state.CB, state.SCB, _ = bnb.functional.int8_vectorwise_quant(B2.to(torch.float16))
+                state.CB, state.SCB, _ = bnb.functional.int8_vectorwise_quant(B2)
                 B2 = state.CB
 
             if not transpose[0] and transpose[1]:
